Open Stage Progression
Adds a new option that forgoes the chapter system in favour of giving each stage its own access item.

This also strips away the Star Card requirements for the stages, which a part of me is debating adding to the Individual Chapters option too.

Also fixed the No Items Chest Tracer not generating (was my own stupidity causing it to remove EVERY location or something, whoops).